day_9/requirements.txt
# PostgreSQL and pgvector dependencies
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
asyncpg==0.29.0

# Embedding models
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing
numpy==1.24.4
pandas==2.1.3

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-postgresql==5.0.0

# Development tools
alembic==1.12.1